+++ projects/sds/src/obliqueVersion.ts
@@ -1,1 +1,1 @@
-export const latest = '11.3.2';
+export const latest = '11.3.3';